Gemfile: Remove `puma` as dependency
The gem is not required to make the service run (fallback is WEBrick).

We remove it because it requires additional native C libraries on the
host and would make the setup more complex without adding value.

We assume that the container is proxied in PROD use cases anyways in
case throughput would be a business goal (which we never saw in real
life as all build screens had relativly low throughput).